Move internal/flag to flag
