+++ flag/BUILD.bazel
@@ -3,6 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["flag.go"],
-    importpath = "github.com/bazelbuild/bazel-gazelle/internal/flag",
+    importpath = "github.com/bazelbuild/bazel-gazelle/flag",
     visibility = ["//:__subpackages__"],
 )
