+++ language/go/BUILD.bazel
@@ -47,7 +47,7 @@ go_library(
     visibility = ["//:__subpackages__"],
     deps = [
         "//config:go_default_library",
-        "//internal/flag:go_default_library",
+        "//flag:go_default_library",
         "//label:go_default_library",
         "//language:go_default_library",
         "//language/proto:go_default_library",
